move release database and base models to separate file
this will prevent issues with circular imports when I add the storage locations model
