+++ backend/app/crud.py
@@ -4,8 +4,9 @@ from typing import Any
 from sqlmodel import Session, select
 
 from app.core.security import get_password_hash, verify_password
+from app.models.database_models import Release
 from app.models.models import Item, ItemCreate, User, UserCreate, UserUpdate
-from app.models.release import Release, ReleaseCreate
+from app.models.release import ReleaseCreate
 
 
 def create_user(*, session: Session, user_create: UserCreate) -> User:
